Slight fix to elevator down animation
Player will transform with the elevator; no longer just falling with the
elevator. However, there is a slight issue with the collider, so the
player will be slightly hovering when the elevator goes down.
Additionally, if the player moves during the elevator down animation, he
will be immediately transported to the right of the elevator. If the
player stands still, the animation will work fine.

After the elevator up animation finishes, the player should no longer be
attached to the elevator transform, and can move freely.